maintained Model controller structure
