+++ api/routes/packages.js
@@ -9,6 +9,7 @@ const getCoordinates = require('../utils/getcoordinates');
 // const User = require('../models/user');
 // const calculateDistance = require('../utils/getcoordinates');
 // GET all packages
+const PackageController = require('../controllers/packages');
 const JWT_KEY = process.env.JWT_KEY || 'your_secret_key';
 const calculateDistance = (coord1, coord2) => {
     const toRadians = (degree) => degree * (Math.PI / 180);
@@ -31,126 +32,17 @@ const calculateDistance = (coord1, coord2) => {
     return R * c; // Distance in kilometers
 };
 
-router.get('/', checkAuth, (req, res, next) => {
-    Package.find()
-        .exec()
-        .then(packages => {
-            res.status(200).json({
-                count: packages.length,
-                packages: packages
-            });
-        })
-        .catch(err => {
-            res.status(500).json({ error: err });
-        });
-});
+router.get('/', checkAuth, PackageController.getAllPackages);
 
 // POST create a new package
-router.post('/', checkAuth, authorize('customer'), async (req, res, next) => {
-    const { pickupLocationName, destinationName, estimatedDeliveryTime, estimatedPrice, deliverTo, userName } = req.body;
-
-    // Fetch coordinates for pickup and destination locations
-    const pickupCoordinates = await getCoordinates(pickupLocationName);
-    const destinationCoordinates = await getCoordinates(destinationName);
-    // console.log(req.userData);
-    console.log(pickupCoordinates, destinationCoordinates)
-    const userId = req.userData.userId;
-    console.log(userId);
-    const newPackage = new Package({
-        _id: new mongoose.Types.ObjectId(),
-        userName,
-        userId,
-        pickupLocationName,
-        pickupCoordinates,
-        destinationName,
-        destinationCoordinates,
-        estimatedDeliveryTime,
-        estimatedPrice,
-        deliverTo,
-        createdOn: new Date(),
-        deliveryDistance: calculateDistance(pickupCoordinates, destinationCoordinates) // Assuming you calculate distance here
-    });
-    console.log(newPackage);
-    await newPackage
-        .save()
-        .then(result => {
-            console.log("hello");
-            // After saving package, add the packageId to the user’s packages array
-            return User.findByIdAndUpdate(
-                req.userData.userId,
-                { $push: { packages: result._id } },  // Push package ID to user's packages array
-                { new: true }
-            ).exec();
-        })
-        .then(user => {
-            res.status(201).json({
-                message: 'Package created successfully',
-                createdPackage: newPackage,
-                user: user  // Updated user with the new package reference
-            });
-        })
-        .catch(err => {
-            console.error("Error occurred:", err.message || err);
-            res.status(500).json({ error: err });
-
-        });
-});
+router.post('/', checkAuth, authorize('customer'),PackageController.createPackage);
 
 // GET a single package by ID
-router.get('/:packageId', checkAuth, (req, res, next) => {
-    const id = req.params.packageId;
-    Package.findById(id)
-        .exec()
-        .then(doc => {
-            if (doc) {
-                res.status(200).json(doc);
-            } else {
-                res.status(404).json({ message: 'No valid entry found for provided ID' });
-            }
-        })
-        .catch(err => {
-            res.status(500).json({ error: err });
-        });
-});
+router.get('/:packageId', checkAuth, PackageController.getPackageById);
 
 // DELETE a package
-router.delete('/', checkAuth, async (req, res, next) => {
-    const userId = req.userData.userId;
-
-    try {
-        // Check if the user exists
-        const user = await User.findById(userId);
-        if (!user) {
-            return res.status(404).json({ message: 'User not found' });
-        }
-
-        // Delete all packages associated with the user
-        const result = await Package.deleteMany({ userId: userId });
-        console.log('Packages Deleted:', result);
-
-        // Optionally, update the user to remove package references
-        await User.findByIdAndUpdate(userId, { $set: { packages: [] } }).exec();
-
-        res.status(200).json({
-            message: 'All packages for the user have been deleted',
-            deletedCount: result.deletedCount
-        });
-    } catch (err) {
-        console.error('Error deleting packages:', err);
-        res.status(500).json({ error: err.message || err });
-    }
-});
-router.delete('/:packageId', checkAuth,authorize('customer'),async (req, res, next) => {
-    const id = req.params.packageId;
-    await Package.deleteOne({ _id: id })
-        .exec()
-        .then(result => {
-            res.status(200).json({ message: 'Package deleted' });
-        })
-        .catch(err => {
-            res.status(500).json({ error: err });
-        });
-});
+router.delete('/', checkAuth, authorize('customer'), PackageController.deletePackagesForUser);
+router.delete('/:packageId', checkAuth,authorize('customer'),PackageController.deletePackageById);
 
 router.post('/near-by', checkAuth,authorize('driver'), async (req, res, next) => {
     console.log("hello");
